// AngularWithWebApi/src/app/Model/ipropertybase.ts
export interface Ipropertybase {
  id?: number;
  sellrent?: number;
  BHK?: number;
  ptype?: string;
  ftype?: string;
  name?: string;
  price?: number;
  builtArea?: number;
  city?: string;
  RTM?: number;
  image?: string;
 
  //id?: number;
  //sellrent?: number;
  //ptype?: string;
  //ftype: string;
  //name: string;
  //price?: number;
  //BHK?: number;
  //builtArea?: number;
  //city: string;
  //RTM?: number;
  //image?: string;
}